requirements.txt
aiosqlite
discord.py
gspread
python-dotenv